Ramp function clean up
We really only call the runtime evaluation function to get the end value
so simply provide an overridable interface for that instead.
The generic rampfunction that compiles python callback is the only one that
needs the old logic and we can simply make this code specific to the generic class.

This removes a bunch of dead code for the specialized implementations.
Also remove now unused import and exports (from other module).